requirements.txt
numpy==1.26.3
Pillow==10.2.0
tensorflow
torch==2.1.2
torchvision==0.16.2
